1CRYPTEN_SPACE_V4.0/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
supabase==2.3.0
python-dotenv==1.0.0
httpx>=0.24.0,<0.25.0
websockets==12.0
google-generativeai==0.3.2
pandas==2.2.0
numpy==1.26.3
pybit==5.7.0
